+++ Measurement.h
@@ -11,7 +11,7 @@
 void initAdc0(void);
 void inintSyncCableInput(void);
 void measurement_signal_generation(void);
-void measurement_measure(unsigned short *);
-void measurement_work(unsigned short *);
+void measurement_measure(uint8_t *);
+void measurement_work(uint8_t *);
 
 #endif /* MEASUREMENT_H_ */
